requirements.txt
langchain
chromadb
pypdf
langchain-mistralai
python-dotenv
unstructured
python-pptx
langchain-community